FSO Part4 ex18-23

+++ controllers/blogs.js
@@ -1,20 +1,24 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog.js')
-const User = require('../models/user.js')
 const logger = require('../utils/logger.js')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({})
-    // Exercise 4.17
     .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
   //logger.info(request.body)
-  const { title, author, url, likes, userId } = request.body
-  const user = await User.findById(userId)
+  const user = request.user
+  logger.info('token user: ', user)
+  const { title, author, url, likes } = request.body
+  //const token = getTokenFrom(request)
+  logger.info('token: ', request.token)
+  if (!user || !request.token) {
+    return response.status(401).json({ error: 'token is missing the first word' })
+  }
   const blog = new Blog({
     title,
     author,
@@ -22,9 +26,8 @@ blogsRouter.post('/', async (request, response) => {
     likes: likes || 0,
     user: user._id
   })
-  logger.info(blog, 'new blog')
-  logger.info(blog.user, 'new blog user id')
-
+  //logger.info(blog, 'new blog')
+  //logger.info(blog.user, 'new blog user id')
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
@@ -36,19 +39,30 @@ blogsRouter.get('/:id', async (request, response) => {
   if (blog) {
     response.json(blog)
   } else {
-    response.status(404).json({ error: 'this blog does not exist' })
+    response.status(404).end()
   }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndRemove(request.params.id)
-  //await Blog.findById(request.params.id)
+  const user = request.user
+  //logger.info('token user: ', user)
+  const blog = await Blog.findById(request.params.id)
   //logger.info('blog info: ', blog)
   //logger.info('blog id info: ', blog.id)  // is a string
   //logger.info('blog user info: ', blog.user.toString())
   if ( !blog ) {
     return response.status(401).json({ error: 'this blog does not exist' })
   }
+  // Exercise 4.21
+  if ( blog.user.toString() !== user.id ){
+    return response.status(401).json({ error: 'you cannot delete nothing by a wrong user' })
+  }
+  //logger.info(user.blogs)
+  //logger.info(user.blogs.filter(blog => blog.toString() !== request.params.id))
+  user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id)
+  await user.save()
+  // Before only was this
+  await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
